feat: implement invalidating tokens

+++ src/graphql/resolvers/Mutations/deleteItem.js
@@ -1,12 +1,16 @@
+import { AuthenticationError } from 'apollo-server';
 import { returnItems } from '../../../database/utils';
-import { userOfListByItemId } from '../checkAuth';
+import { userOfListByItemId, isTokenValid } from '../checkAuth';
 import { ITEM_DELETED } from '../../events';
 
 export const deleteItem = async (
   root,
   { itemId },
-  { models: { Item, User, List }, user, pubsub },
+  { models: { Item, User, List, BlacklistedToken }, user, pubsub, token },
 ) => {
+  if (!(await isTokenValid(token, BlacklistedToken))) {
+    throw new AuthenticationError('Invalid token');
+  }
   if (userOfListByItemId(user, itemId, User, List) || user.isAdmin) {
     const deletedItem = returnItems(await Item.findByIdAndRemove(itemId));
     pubsub.publish(ITEM_DELETED, {
